requirements.txt
selenium==4.15.2
opencv-python==4.8.1.78
pytesseract==0.3.10
python-telegram-bot==21.7
undetected-chromedriver==3.5.8
webdriver-manager==4.0.1
numpy==1.24.0
Pillow==10.0.0
